% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_westbay.R
\name{read_westbay.character}
\alias{read_westbay.character}
\title{read_westbay.character}
\usage{
\method{read_westbay}{character}(x, transducer_depth = NULL, well_elevation = NULL, ...)
}
\arguments{
\item{x}{name of the file}

\item{...}{}
}
\description{
read_westbay.character
}
\examples{
#===============================================================================
}
