% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_westbay.R
\name{read_westbay}
\alias{read_westbay}
\title{read_westbay}
\usage{
read_westbay(x, transducer_depth = NULL, well_elevation = NULL, ...)
}
\arguments{
\item{x}{name of the file}

\item{...}{}
}
\description{
read_westbay
}
